Fix renderinf og OperationEditor
